ml/requirements.txt
numpy
matplotlib
pillow
scikit-learn
tensorflow

